+++ src/components/mypagefriend/follower.tsx
@@ -1,25 +1,6 @@
 import RoundBtn from "../layout/roundBtn";
 import React, { useState } from "react";
-
-interface FriendProps {
-    name: string;
-}
-
-function Friend (props: FriendProps) {
-    // 친구 프사 + 닉네임 + 버튼
-    return (
-        <div className="flex my-4 justify-between">
-            <div className="flex items-center">
-                <img className="mr-5" src="/images/user.svg" alt="친구 프로필사진" />
-                {props.name}
-            </div>
-            <div className="flex">
-                <RoundBtn label="수락" color="bg-lightgrey" px={10}/>
-                <RoundBtn label="거절" color="bg-lightgrey" px={10}/>
-            </div>
-        </div>
-    )
-}
+import FriendTwoBtn from "./friendTwoBtn";
 
 function Follower () {
     const [friends, setFriends] = useState([
@@ -44,7 +25,7 @@ function Follower () {
             </div>
             <div className="mt-12">
                 {friends.map((friend, index) => (
-                    <Friend key={index} name={friend} />
+                    <FriendTwoBtn key={index} name={friend} label1="수락" label2="거절" px={6}/>
                 ))}
             </div>
         </div>
