+++ src/components/mypagefriend/following.tsx
@@ -1,24 +1,6 @@
 import RoundBtn from "../layout/roundBtn";
 import React, { useState } from "react";
-
-interface FriendProps {
-    name: string;
-}
-
-function Friend (props: FriendProps) {
-    // 친구 프사 + 닉네임 + 버튼
-    return (
-        <div className="flex my-4 justify-between">
-            <div className="flex items-center">
-                <img className="mr-5" src="/images/user.svg" alt="친구 프로필사진" />
-                {props.name}
-            </div>
-            <div className="flex">
-                <RoundBtn label="요청취소" color="bg-lightgrey" px={6}/>
-            </div>
-        </div>
-    )
-}
+import FriendOneBtn from "./friendOneBtn";
 
 function Following () {
     const [friends, setFriends] = useState([
@@ -44,7 +26,7 @@ function Following () {
             </div>
             <div className="mt-12">
                 {friends.map((friend, index) => (
-                    <Friend key={index} name={friend} />
+                    <FriendOneBtn key={index} name={friend} label1="요청취소" px={6}/>
                 ))}
             </div>
         </div>
