+++ src/components/mypagefriend/sendFollow.tsx
@@ -1,26 +1,7 @@
 import { BiSearch } from "react-icons/bi";
 import RoundBtn from "../layout/roundBtn";
 import React, { useState } from "react";
-
-interface FriendProps {
-    name: string;
-}
-
-function Friend (props: FriendProps) {
-    // 친구 프사 + 닉네임 + 버튼
-    return (
-        <div className="flex my-4 justify-between">
-            <div className="flex items-center">
-                <img className="mr-5" src="/images/user.svg" alt="친구 프로필사진" />
-                {props.name}
-            </div>
-            <div className="flex">
-                <RoundBtn label="요청" color="bg-lightgrey" px={10} />
-                <RoundBtn label="차단" color="bg-lightgrey" px={10} />
-            </div>
-        </div>
-    )
-}
+import FriendTwoBtn from "./friendTwoBtn";
 
 function SendFollow () {
     const [friends, setFriends] = useState([
@@ -54,7 +35,7 @@ function SendFollow () {
                 </div>
                 <div className="mt-6">
                     {friends.map((friend, index) => (
-                        <Friend key={index} name={friend} />
+                        <FriendTwoBtn key={index} name={friend} label1="요청" label2="차단" px={6} />
                     ))}
                 </div>
             </div>
